+++ src/experimental/decorators/eip5792.test.ts
@@ -26,7 +26,7 @@ const client = createClient({
           },
         }
       if (method === 'wallet_sendCalls') return '0x1'
-      if (method === 'wallet_getCallsReceipt')
+      if (method === 'wallet_getCallsStatus')
         return {
           status: 'CONFIRMED',
           receipts: [
@@ -50,7 +50,7 @@ const client = createClient({
 test('default', async () => {
   expect(walletActionsEip5792()(client)).toMatchInlineSnapshot(`
     {
-      "getCallsReceipt": [Function],
+      "getCallsStatus": [Function],
       "getCapabilities": [Function],
       "sendCalls": [Function],
       "writeContracts": [Function],
@@ -60,7 +60,11 @@ test('default', async () => {
 
 describe('smoke test', () => {
   test('getCapabilities', async () => {
-    expect(await client.getCapabilities()).toMatchInlineSnapshot(`
+    expect(
+      await client.getCapabilities({
+        account: '0x0000000000000000000000000000000000000000',
+      }),
+    ).toMatchInlineSnapshot(`
       {
         "8453": {
           "paymasterService": {
@@ -79,10 +83,8 @@ describe('smoke test', () => {
     `)
   })
 
-  test('getCallsReceipt', async () => {
-    expect(
-      await client.getCallsReceipt({ id: '0x123' }),
-    ).toMatchInlineSnapshot(`
+  test('getCallsStatus', async () => {
+    expect(await client.getCallsStatus({ id: '0x123' })).toMatchInlineSnapshot(`
       {
         "receipts": [
           {